[fix] 어플리케이션 배포 단위를 서비스 및 배포 환경 단위로 수정

+++ argo/Jenkinsfile-application.groovy
@@ -2,7 +2,8 @@ pipeline {
     agent any
 
     parameters {
-        choice(choices: ['option', 'comppi-dev', 'comppi-prod', 'comppi-gateway'], name: 'APPLICATION', description: '어플리케이션 선택')
+        choice(choices: ['option', 'gateway', 'api-gateway', 'auth-service', 'board-service', 'encyclo-service', 'my-plant', 'minio', 'redis', 'discovery-eureka'], name: 'APPLICATION', description: '어플리케이션 선택')
+        choice(choices: ['option', 'dev', 'prod'], name: 'ENVIRONMENT', description: '배포 환경')
         choice(choices: ['option', 'upgrade', 'uninstall'], name: 'DEPLOY_TYPE', description: '배포/삭제')
     }
 
@@ -18,12 +19,22 @@ pipeline {
             steps {
                 script {
                     if (params.DEPLOY_TYPE == "upgrade") {
-                        HELM_DEPLOY_COMMAND = "helm upgrade ${params.APPLICATION} ${ARGO_PATH}/${params.APPLICATION} " +
-                                " -f ${ARGO_PATH}/${params.APPLICATION}/values.yaml " +
-                                " -n ${NAMESPACE} --install --wait --timeout=10m "
+                        if (params.APPLICATION == "gateway") {
+                            HELM_DEPLOY_COMMAND = "helm upgrade ${params.APPLICATION} ${ARGO_PATH}/${params.APPLICATION} " +
+                                    " -f ${ARGO_PATH}/${params.APPLICATION}/values.yaml " +
+                                    " -n ${NAMESPACE} --install --wait --timeout=10m "
                             sh "eval ${HELM_DEPLOY_COMMAND}"
+                        } else {
+                            HELM_DEPLOY_COMMAND = "helm upgrade ${params.APPLICATION}-${params.ENVIRONMENT} ${ARGO_PATH}/${params.APPLICATION} " +
+                                    " -f ${ARGO_PATH}/${params.APPLICATION}/values-${params.ENVIRONMENT}.yaml " +
+                                    " -n ${NAMESPACE} --install --wait --timeout=10m "
+                            sh "eval ${HELM_DEPLOY_COMMAND}"
+                        }
                     } else if (params.DEPLOY_TYPE == "uninstall") {
-                        sh "helm uninstall ${params.APPLICATION} -n argo"
+                        if (params.APPLICATION == "gateway"){
+                            sh "helm uninstall ${params.APPLICATION} -n argo"
+                        }
+                            sh "helm uninstall ${params.APPLICATION}-${params.ENVIRONMENT} -n argo"
                     } else {
                         echo "skip deploy"
                     }
